src/requirements.txt
fastapi
uvicorn
python-decouple
jwt
python-socketio
pandas
openpyxl
